# Add niche optimization 