This PR adds niche optimization. This lets the compiler use values outside of the valid integer range to store e.g. other enum variants. This makes `Option<u7>` one byte.

The internal representation is own the same as the related primitive type, i.e. every u30 can be transmuted to u32. The other direction is possible if the u32 is representable in the u30 range.