@Kijewski 2022-09-14
> Do you know if we have managed to not expose any of the underlying implementation details such that this can be considered backwards compatible?

I code it fully self-contained. The API was not changed in any way, nothing added, nothing removed.

But any code that relied on the previous, undocumented behavior, that you could set the unused bits to any value, will break. This could only be done on `unsafe` code using type casts, so I don't think that matters.

> Would it make sense to add some sort of test to verify that the niche optimization is being made when possible?

[static_assertions](https://crates.io/crates/static_assertions) could be used to ensure that the optimizations work. I'll have a look what the best way is to add it.

> I will try to check out the actual build time one of the next days

I added a commit to this PR, that makes every method `#[inline]` and cuts the compilation time in half: https://github.com/kjetilkjeka/uX/pull/49/commits/1c7a8efe3ca8e217c868ef41655734f73cf07a6b. (It should probably be it's own PR, though.)