@ironhaven 2022-09-16
So I just finished testing how this pull affects compile times and here are the results. The test was I ran `cargo check --timings` on the hello world binary crate with this library as a git dependancy and took the time to check the library alone.

|Computer       |Master branch |Pull    branch | Percentage|
|---|---|---| ---| 
|Win 10 Ryzen 3600x (6 core 4Ghz)        |10.09s |11.02s|+9.21%|
|Manjaro arm Pinebook Pro (2 cores 2Ghz, 4 cores 1.5Ghz)|57.12s |63.6s  | +1134%|

[test data](https://github.com/kjetilkjeka/uX/files/9581351/tests.zip)

My testing shows this pull will increase compile time by about 10%. That is the raw data.