@Kijewski 2022-09-16
I optimized the macro invocation, and used the same test as @ironhaven (best of three runs)

master | pr | percentage
----|---|---
8.75s | 8.87s | +1.4%

cargo 1.63.0 (fd9c4297c 2022-07-01), AMD Ryzen 7 2700X (8 cores @ 3.7GHz)

The optimization could easily be used without without the adding the niche optimization, so it would be a lie to claim this PR comes without a cost.