@kjetilkjeka 2022-09-23
@chrysn do you think you will have a chance to test it out before merge? Or should I try to move this on without?