@chrysn 2022-09-23
I've attempted a larger trial run, but that got stuck in too many places where I'd need to introduce additional `.into()` calls to make things practical. (I still plan to use it there, but it'll have to wait).

I've done a few tests more on the micro side, finding that
* `Result<u31, ()>` and other ZSTs (like single-inhabitant repr(Rust) enums) are good / small (4 byte),
* `Result<u31, T>` with a multi-valued enums are large (8 byte),
* and even `Result<u31, T>` with a `repr(u32)` enums with a value in which the MSB-set is large (8 byte).

That doesn't make me particularly happy, but it's still an improvement, and most importantly, to my understanding, further improvements would not need to be done here, but would be up to the Rust compiler.

So I have no clear recommendation -- it's definitely not a "no", but it's not the enthusiastic "yes" I hoped to provide either. Given you were about to merge even without my tests, I'd guess that's a "go ahead".