@Kijewski 2022-09-23
@chrysn, that should be fixed by https://github.com/rust-lang/rust/pull/94075. I don't know when this change lands in `stable`.