@chrysn 2022-09-23
I should have tested on nightly right away -- a Result<u31, multi-valued-enum> is apparently optimized starting with current beta, 1.65-to-be. What it still can't do is optimizing for enums with values (eg. `enum Error { Val1 = 0xffffffff, Val2 = 0xff00ffff }`), like we'd have if we introduced a i32neg type that represents the inclusive range from -1 to -2**32 (where `Result<u31, i32neg>` would be practically u32 but treatable like a result).

So ... well, one step closer even. Thanks, I like it :-)