@dacut 2023-01-17
I was trying to see what was still needed for this PR. Seems like it's just build times that are a concern?

Here's the latest on my M1 MacBook Pro, if this is of any help (hello world cargo binary using uX as a dep, with just enough code so it's not elided, `cargo clean; cargo check --timings`):

* master: Min 5.72 s/Mean 5.78 s/Max 5.93 s
* pr-niche: Min 5.79 s/Mean 5.94 s/Max 6.05 s

It's a 3% gain, but with some overlap in the data.